asset table and prevent flickering of views

+++ Gaming/Stocks/pattern_database/run_daily_tasks.py
@@ -14,33 +14,38 @@ stock_db_updater = StockDatabaseUpdater()
 pattern_table = PatternTable()
 trade_table = TradeTable()
 
-is_weekday = datetime.today().weekday() in [1, 2, 3, 4, 5]  # tuesday till saturday => update stock data as well
+process_single_ticker_id = False
 
-stock_db.delete_duplicate_records(trade_table)
-stock_db.delete_duplicate_records(pattern_table)
-stock_db.update_crypto_currencies(PRD.DAILY)
-#
-if is_weekday:
-    stock_db.update_stock_data_by_index(INDICES.DOW_JONES, PRD.DAILY)
-    stock_db.update_stock_data_for_symbol('FCEL')
-    stock_db.update_stock_data_for_symbol('TSLA')
-    stock_db.update_stock_data_for_symbol('GE')
+if process_single_ticker_id:
+    stock_db_updater.update_wave_records_for_daily_period('AAPL', 400)
+else:
+    is_weekday = datetime.today().weekday() in [1, 2, 3, 4, 5]  # tuesday till saturday => update stock data as well
+
+    stock_db.delete_duplicate_records(trade_table)
+    stock_db.delete_duplicate_records(pattern_table)
+    stock_db.update_crypto_currencies(PRD.DAILY)
     #
-    stock_db.update_stock_data_by_index(INDICES.NASDAQ100, PRD.DAILY)
+    if is_weekday:
+        stock_db.update_stock_data_by_index(INDICES.DOW_JONES, PRD.DAILY)
+        stock_db.update_stock_data_for_symbol('FCEL')
+        stock_db.update_stock_data_for_symbol('TSLA')
+        stock_db.update_stock_data_for_symbol('GE')
+        #
+        stock_db.update_stock_data_by_index(INDICES.NASDAQ100, PRD.DAILY)
 
-# print(stock_db.get_wave_counter_dict(400))
+    # print(stock_db.get_wave_counter_dict(400))
 
-stock_db_updater.update_wave_data_by_index_for_daily_period(INDICES.CRYPTO_CCY, 400)
-if is_weekday:
-    stock_db_updater.update_wave_data_by_index_for_daily_period(INDICES.DOW_JONES, 400)
-    stock_db_updater.update_wave_data_by_index_for_daily_period(INDICES.NASDAQ100, 400)
+    stock_db_updater.update_wave_data_by_index_for_daily_period(INDICES.CRYPTO_CCY, 400)
+    if is_weekday:
+        stock_db_updater.update_wave_data_by_index_for_daily_period(INDICES.DOW_JONES, 400)
+        stock_db_updater.update_wave_data_by_index_for_daily_period(INDICES.NASDAQ100, 400)
 
-stock_db_updater.update_wave_data_by_index_for_intraday(INDICES.CRYPTO_CCY)
-if is_weekday:
-    stock_db_updater.update_wave_data_by_index_for_intraday(INDICES.DOW_JONES)
-    stock_db_updater.update_wave_data_by_index_for_intraday(INDICES.NASDAQ100)
+    stock_db_updater.update_wave_data_by_index_for_intraday(INDICES.CRYPTO_CCY)
+    if is_weekday:
+        stock_db_updater.update_wave_data_by_index_for_intraday(INDICES.DOW_JONES)
+        stock_db_updater.update_wave_data_by_index_for_intraday(INDICES.NASDAQ100)
 
-if is_weekday:
-    stock_db_updater.update_trade_records(16, 16)
+    if is_weekday:
+        stock_db_updater.update_trade_records(16, 16)
